+++ tests/neuralnet/test_network_evaluation_complex.py
@@ -1,4 +1,4 @@
-from decimal import Decimal
+
 import math
 import unittest
 from src.neuralnet.network import Network
@@ -21,15 +21,15 @@ class TestNetworkEvaluationComplex(unittest.TestCase):
         
         
     def validate_layer_1_activation(self, network: Network):
-        is_close_assert(network.node_layers[0][0].activation, Decimal('0.9706'), abs_tol=0.001)
-        is_close_assert(network.node_layers[0][1].activation, Decimal('0.9836'), abs_tol=0.001)
+        is_close_assert(network.node_layers[0][0].activation, float('0.9706'), abs_tol=0.001)
+        is_close_assert(network.node_layers[0][1].activation, float('0.9836'), abs_tol=0.001)
         
         
     def validate_layer_2_activation(self, network: Network):
-        is_close_assert(network.node_layers[1][0].activation, Decimal('0.9642'), abs_tol=0.001)
-        is_close_assert(network.node_layers[1][1].activation, Decimal('0.8784'), abs_tol=0.001)
-        is_close_assert(network.node_layers[1][2].activation, Decimal('0.9407'), abs_tol=0.001)
+        is_close_assert(network.node_layers[1][0].activation, float('0.9642'), abs_tol=0.001)
+        is_close_assert(network.node_layers[1][1].activation, float('0.8784'), abs_tol=0.001)
+        is_close_assert(network.node_layers[1][2].activation, float('0.9407'), abs_tol=0.001)
     
     def validate_layer_3_activation(self, network: Network):
-        is_close_assert(network.node_layers[2][0].activation, Decimal('0.9586'), abs_tol=0.001)
-        is_close_assert(network.node_layers[2][1].activation, Decimal('0.9715'), abs_tol=0.001)
+        is_close_assert(network.node_layers[2][0].activation, float('0.9586'), abs_tol=0.001)
+        is_close_assert(network.node_layers[2][1].activation, float('0.9715'), abs_tol=0.001)
